refactor: factor out getLanguage listener
Factor out the getLanguage listener to its own file.

refs #104# Please enter the commit message for your changes. Lines starting
